Refactor: parser.test.js -> parser-util.test.js
